[Enhance] refactor todo-list - completeHandler 추가
